requirements.txt
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
attrs==25.3.0
Brotli==1.1.0
cffi==1.17.1
frozenlist==1.7.0
idna==3.10
msgspec==0.18.6
multidict==6.6.1
propcache==0.3.2
py-cord==2.6.1
pycares==4.9.0
pycparser==2.22
python-dotenv==1.1.1
yarl==1.20.1
aiodns==3.5.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
frozenlist==1.7.0
google-auth==2.40.3
google-genai==1.23.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
msgspec==0.18.6
multidict==6.6.1
propcache==0.3.2
py-cord==2.6.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycares==4.9.0
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
requests==2.32.4
rsa==4.9.1
sniffio==1.3.1
soupsieve==2.7
tenacity==8.5.0
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
